src/requirements.txt
attr==0.3.2
ConfigParser==5.3.0
cryptography==41.0.1
Cython==0.29.35
dl==0.1.0
docutils==0.20.1
Flask==2.3.2
Flask_Bcrypt==1.0.1
Flask_JWT_Extended==4.5.2
flask_marshmallow==0.15.0
flask_sqlalchemy==3.0.5
HTMLParser==0.0.2
ipython==8.14.0
ipywidgets==8.0.6
Jinja2==3.1.2
jnius==1.1.0
keyring==24.2.0
lockfile==0.12.2
marshmallow==3.19.0
numpy==1.24.3
Pillow==10.0.0
protobuf==4.23.3
pyOpenSSL==23.2.0
python-dotenv==1.0.0
railroad==0.5.0
Sphinx==7.0.1
SQLAlchemy==2.0.15
tornado==6.3.2
trove_classifiers==2023.5.24
truststore==0.7.0
urllib3_secure_extra==0.1.0
Werkzeug==2.3.4
xmlrpclib==1.0.1
